feat: make it so only users who created the blog can delete it

+++ blog_list/controllers/blogs.js
@@ -2,7 +2,11 @@ import Blog from "../models/blog.js";
 import User from "../models/user.js";
 
 async function getAllBlogs(req, res) {
-    let allBlogs = await Blog.find({}).populate("user", {username: 1, name: 1, id: 1});
+    let allBlogs = await Blog.find({}).populate("user", {
+        username: 1,
+        name: 1,
+        id: 1,
+    });
     allBlogs = allBlogs.map((blog) => blog.toJSON());
 
     res.json(allBlogs);
@@ -11,9 +15,9 @@ async function getAllBlogs(req, res) {
 async function postNewBlog(req, res) {
     const user = req.user;
 
-    const newBlog = new Blog({...req.body, user: user._id});
+    const newBlog = new Blog({ ...req.body, user: user._id });
     const saveResult = await newBlog.save();
-    user.blogs = user.blogs.concat(saveResult._id)
+    user.blogs = user.blogs.concat(saveResult._id);
     await user.save();
 
     res.status(201).json(saveResult.toJSON());
@@ -21,6 +25,16 @@ async function postNewBlog(req, res) {
 
 async function deleteBlogById(req, res) {
     const { id } = req.params;
+
+    const user = req.user;
+
+    const userContainsBlogId = user.blogs.includes(id);
+
+    if (!userContainsBlogId) {
+        return res
+            .status(403)
+            .json({ error: "you are not authorized to delete this blog" });
+    }
 
     const deletedBlog = await Blog.findByIdAndDelete(id);
 
@@ -33,7 +47,7 @@ async function deleteBlogById(req, res) {
 
 async function updateBlogById(req, res) {
     const { id } = req.params;
-    const  body  = req.body;
+    const body = req.body;
 
     const blog = await Blog.findById(id);
